Prompt user for permission before replacing files required for install

+++ install.sh
@@ -11,13 +11,48 @@ fatal()   { echo "[FATAL]   $@" | tee -a "$LOG_FILE" >&2 ; exit 1 ; }
 
 wget https://github.com/eddyerburgh/git-init-plus/archive/master.zip
 
-info "Downloading zip from github"
+# Prompt for permission to remove any existing files/folders that are used in install
+if [ -e "/opt/git-init-plus-master.zip" ]; then
+  read -r -p "/opt/git-init-plus-master.zip already exists, OK to replace? [y/N] " response
+  case "$response" in
+    [yY][eE][sS]|[yY])
+      rm /opt/git-init-plus-master.zip
+      mv master.zip /opt/git-init-plus-master.zip
+      ;;
+    *)
+      info "Install exiting. Please rename /opt/git-init-plus-master.zip and rerun install script"
+      exit 2
+    ;;
+  esac
+else
+  mv master.zip /opt/git-init-plus-master.zip
+fi
 
-[ -e /opt/git-init-master.zip ] && rm /opt/git-init-master.zip
-[ -e /opt/git-init-plus ] && rm -rf /opt/git-init-plus
-[ -e /opt/git-init-plus-master ] && rm -rf /opt/git-init-plus-master
+if [ -e "/opt/git-init-plus-master" ]; then
+  read -r -p "/opt/git-init-plus-master directory already exists, OK to replace? [y/N] " response
+  case "$response" in
+    [yY][eE][sS]|[yY])
+      rm -rf "/opt/git-init-plus-master"
+      ;;
+    *)
+      info "Install exiting. Please rename /opt/git-init-plus-master and rerun install script"
+      exit 2
+    ;;
+  esac
+fi
 
-mv master.zip /opt/git-init-plus-master.zip
+if [ -e "/opt/git-init-plus" ]; then
+  read -r -p "/opt/git-init-plus directory already exists, OK to replace? [y/N] " response
+  case "$response" in
+    [yY][eE][sS]|[yY])
+      rm -rf "/opt/git-init-plus"
+      ;;
+    *)
+      info "Install exiting. Please rename /opt/git-init-plus and rerun install script"
+      exit 2
+    ;;
+  esac
+fi
 
 cd /opt
 unzip git-init-plus-master.zip
@@ -32,5 +67,5 @@ ln -sf "$ROOT_DIR/git-init-plus.sh" /usr/local/bin/git-init-plus
 
 info "Success! git-init-plus is installed in /usr/local/bin
 Run git-init-plus to start a new git project
-If git-init-plus is no recognised as a command, add /usr/loca/bin to your PATH
+If git-init-plus is no recognised as a command, add /usr/local/bin to your PATH
 "
